requirements.txt
amqp>=5.0.6
asgiref>=3.4.1
billiard>=3.6.4.0
celery>=5.2.1
celery-progress>=0.1.1
certifi>=2021.10.8
charset-normalizer>=2.0.7
click>=8.0.3
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.2.0
Deprecated>=1.2.13
Django>=3.2.9
django-celery-beat>=2.2.1
django-cors-headers>=3.11.0
django-mysql>=3.8.1
django-timezone-field>=4.2.1
djangorestframework>=3.13.1
et-xmlfile>=1.1.0
greenlet>=1.1.2
gunicorn>=20.1.0
idna>=3.3
kombu>=5.2.2
lxml>=4.6.4
mysql-connector-python>=8.0.26
numpy>=1.21.4
openpyxl>=3.0.9
pandas>=1.3.3
prompt-toolkit>=3.0.22
protobuf>=3.19.1
psycopg2>=2.9.3
PyMySQL>=1.0.2
python-crontab>=2.6.0
python-dateutil>=2.8.2
python-dotenv>=0.20.0
python-environ>=0.4.54
pytz>=2021.3
redis>=4.0.2
requests>=2.26.0
six>=1.16.0
SQLAlchemy>=1.4.25
sqlparse>=0.4.2
urllib3>=1.26.7
vine>=5.0.0
wcwidth>=0.2.5
wrapt>=1.13.3
